Added Node class implementation
The Node class is designed to create list items. An instance of a class is an object that has 3 fields: a value, a reference to the next item in the list, and a reference to the previous item in the list. The node value can only be a string of length 1. The Node class will be used in the implementation of the List class.