Added dlList class and length method
The class will consist of three fields and methods that will be developed later. The head and tail of the list will store links to the first and last nodes of the list, and the length will indicate the number of nodes. These 3 fields will be encapsulated, so you will need the length method to find out the length of the list