Added append() and insert() methods
The append() method takes a character and adds an element with that value to the end of the list. An error will be thrown when trying to add a non-string element or a string longer than 2 characters. The insert() method is designed to add the specified character to the specified position by shifting other items in the list. An error will be thrown if you specify an item index less than 0 or greater than the length of the list.