Added findFirst() and findLast() methods
Method findFirst() takes the value of the item as an argument and tries to find the first item in the list with that value. Method findLast() tries to find the last element with the specified value. If there is no element with this value in the list, the method returns -1.